requirements.txt
Django==2.2.19
social-auth-app-django==4.0.0
djangorestframework==3.7.3
django-widget-tweaks==1.4.1
Pillow==8.0.0
django-friendship==1.9.1
django-multiselectfield==0.1.12